Fix DPS build script to install correct maap-py version
Since conda lock files purposely do not handle source dependencies, the
DPS build script must install maap-py (from sources) separately from
dependencies specified in the lock file, until maap-py is properly
published to PyPI or conda-forge.

The pip source reference to maap-py was repeated in 3 files, but the
previous patch version of the gedi-subset algo only updated 2 of the 3
files, having missed the build.sh script. Therefore, this patch removes
the duplication in order to completely avoid such a miss in the future.